Do the deployment part of sync
