+++ src/commands/sync-from-workspace.ts
@@ -5,25 +5,31 @@ import { listPolicies, ConfigMap, policyIsDevRego } from '../opa';
 import { failed, Errorable, Failed, succeeded } from '../utils/errorable';
 import { partition } from '../utils/array';
 import { basename } from 'path';
+import { DeploymentInfo, createOrUpdateConfigMapFrom } from '../opa/deployment';
 
 export async function syncFromWorkspace() {
+    const clusterExplorer = await k8s.extension.clusterExplorer.v1;
     const kubectl = await k8s.extension.kubectl.v1;
-    if (!kubectl.available) {
+    if (!clusterExplorer.available) {
+        await showUnavailable(clusterExplorer.reason);
+        return;
+    } else if (!kubectl.available) {
         await showUnavailable(kubectl.reason);
         return;
     }
 
-    await trySyncFromWorkspace(kubectl.api);
+    await trySyncFromWorkspace(clusterExplorer.api, kubectl.api);
 }
 
-async function trySyncFromWorkspace(kubectl: k8s.KubectlV1): Promise<void> {
+async function trySyncFromWorkspace(clusterExplorer: k8s.ClusterExplorerV1, kubectl: k8s.KubectlV1): Promise<void> {
     // Strategy:
     // * Find all the .rego files in the workspace
     // * Look at all the configmaps in the cluster (except system ones)
     // * Deploy all the .rego files EXCEPT those that would overwrite a NON-MANAGED configmap
     //   * TODO: OPTIMISATION: Skip .rego files where the data already matches the .rego file content
     // * Delete all the MANAGED configmaps that do NOT correspond to any .rego file
-    // TO CONSIDER: List what we are going to do first...
+
+    // TODO: We need to make sure all .rego files are saved (per the Deploy command)
 
     const plan = await longRunning('Working out sync actions...', () => syncActions(kubectl));
     if (failed(plan)) {
@@ -46,24 +52,27 @@ async function trySyncFromWorkspace(kubectl: k8s.KubectlV1): Promise<void> {
         return;
     }
 
-    const selectedActions = selectedActionQuickPicks.map((a) => ({ value: a.value, action: a.action }));
-    const selectedActionPromises = selectedActions.map((a) => runAction(kubectl, a));
+const selectedActionPromises = selectedActionQuickPicks.map((a) => runAction(kubectl, a));
     const actionResults = await longRunning('Syncing the cluster from the workspace...', () =>
         Promise.all(selectedActionPromises)
     );
 
     const failures = actionResults.filter((r) => failed(r)) as Failed[];
+    const successCount = actionResults.filter((r) => succeeded(r)).length;
     if (failures.length > 0) {
-        const successCount = actionResults.filter((r) => succeeded(r)).length;
         const successCountInfo = successCount > 0 ? `.  (${successCount} other update(s) succeeded.)` : '';
         await vscode.window.showErrorMessage(`${failures.length} update(s) failed: ${failures.map((f) => f.error[0]).join(', ')}${successCountInfo}`);
         return;
     }
 
+    if (successCount > 0) {
+        clusterExplorer.refresh();
+    }
+
     await vscode.window.showInformationMessage(`Synced the cluster from the workspace`);
 }
 
-function runAction(kubectl: k8s.KubectlV1, action: { value: string, action: 'deploy' | 'delete' }): Promise<Errorable<null>> {
+function runAction(kubectl: k8s.KubectlV1, action: ActionQuickPickItem): Promise<Errorable<null>> {
     switch (action.action) {
         case 'deploy': return runDeployAction(kubectl, action.value);
         case 'delete': return runDeleteAction(kubectl, action.value);
@@ -71,16 +80,19 @@ function runAction(kubectl: k8s.KubectlV1, action: { value: string, action: 'dep
 }
 
 interface SyncActions {
-    readonly deploy: ReadonlyArray<string>;
-    readonly overwriteDevRego: ReadonlyArray<string>;
-    readonly overwriteNonDevRego: ReadonlyArray<string>;
+    readonly deploy: ReadonlyArray<vscode.Uri>;
+    readonly overwriteDevRego: ReadonlyArray<vscode.Uri>;
+    readonly overwriteNonDevRego: ReadonlyArray<vscode.Uri>;
     readonly delete: ReadonlyArray<string>;
 }
 
-interface ActionQuickPickItem extends vscode.QuickPickItem {
+type ActionQuickPickItem = vscode.QuickPickItem & ({
+    readonly value: vscode.Uri;
+    readonly action: 'deploy';
+} | {
     readonly value: string;
-    readonly action: 'deploy' | 'delete';
-}
+    readonly action: 'delete';
+});
 
 async function syncActions(kubectl: k8s.KubectlV1): Promise<Errorable<SyncActions>> {
     const regoUris = await vscode.workspace.findFiles('**/*.rego');
@@ -98,7 +110,7 @@ async function syncActions(kubectl: k8s.KubectlV1): Promise<Errorable<SyncAction
 
     const localRegoFiles = regoUris.map((u) => vscode.workspace.asRelativePath(u));
 
-    const fileActions = partition(localRegoFiles, (f) => deploymentAction(clusterPolicies.result, f));
+    const fileActions = partition(regoUris, (f) => deploymentAction(clusterPolicies.result, f));
     const filesToDeploy = fileActions.get('no-overwrite') || [];
     const filesOverwritingDevRego = fileActions.get('overwrite-dev') || [];
     const filesOverwritingNonDevRego = fileActions.get('overwrite-nondev') || [];
@@ -117,8 +129,8 @@ async function syncActions(kubectl: k8s.KubectlV1): Promise<Errorable<SyncAction
     };
 }
 
-function deploymentAction(policies: ReadonlyArray<ConfigMap>, regoFilePath: string): 'no-overwrite' | 'overwrite-dev' | 'overwrite-nondev' {
-    const policyName = basename(regoFilePath, '.rego');
+function deploymentAction(policies: ReadonlyArray<ConfigMap>, regoFileUri: vscode.Uri): 'no-overwrite' | 'overwrite-dev' | 'overwrite-nondev' {
+    const policyName = basename(regoFileUri.fsPath, '.rego');  // TODO: deduplicate - seems like DeploymentInfo might do this for us?
     const matchingPolicy = policies.find((p) => p.metadata.name === policyName);
     if (!matchingPolicy) {
         return 'no-overwrite';
@@ -130,8 +142,11 @@ function hasMatchingRegoFile(regoFiles: ReadonlyArray<string>, policy: ConfigMap
     return regoFiles.some((f) => basename(f, '.rego') === policy.metadata.name);
 }
 
-async function runDeployAction(kubectl: k8s.KubectlV1, regoFilePath: string): Promise<Errorable<null>> {
-    return { succeeded: true, result: null };
+async function runDeployAction(kubectl: k8s.KubectlV1, regoFileUri: vscode.Uri): Promise<Errorable<null>> {
+    const regoFilePath = regoFileUri.fsPath;
+    const regoFileContent = (await vscode.workspace.openTextDocument(regoFileUri)).getText();
+    const deploymentInfo = new DeploymentInfo(regoFilePath, regoFileContent);
+    return await createOrUpdateConfigMapFrom(deploymentInfo, kubectl);
 }
 
 async function runDeleteAction(kubectl: k8s.KubectlV1, policyName: string): Promise<Errorable<null>> {
